wireguard_manager/requirements.txt
fastapi[all]
pydantic
uvicorn
sqlalchemy
databases
pynacl
python-jose[cryptography]
